#96 Start writing NavigationMap, holder for navigation data on macro and micro level, to speed-up pathfinding. Currently implemented NavigationCell for store navigation statistic of MicroCell, so Pathfinder can create smart path between cells. I have used average speed modifier, average blocks normal and roughness stat (standard deviation of blocks normal). The weights of this coefficients need to be tweaked.

+++ Assets/Scripts/Macro/Coord.cs
@@ -51,7 +51,7 @@ namespace TerrainDemo.Macro
 
         public override int GetHashCode()
         {
-            return X | (Z << 16);
+            return (int)X | ((int)Z << 16);
         }
 
         public override bool Equals(object obj)
